requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
pytz==2025.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Pillow==10.0.0
django-imagekit==5.0.0 